LLVM.General.Analysis

Description

functionality for analyzing Modules. Much of the analysis possible with LLVM is managed internally, as needed by Transforms, and so is not yet exposed here.

Documentation

verify :: Module -> ErrorT String IO ()

Run basic sanity checks on a Module. Note that the same checks will trigger assertions within LLVM if LLVM was built with them turned on, before this function can be is called.