LLVM.General.CommandLine

Description

Tools for processing command line arguments, for command line tools build with llvm (or for other uses forced into pretending to be such to get at (ack) global state).

Documentation

parseCommandLineOptions :: [String] -> Maybe String -> IO ()

http://llvm.org/doxygen/namespacellvm_1_1cl.html#a992a39dae9eb8d4e54ffee5467902803 Sadly, there is occasionally some configuration one would like to control in LLVM which are accessible only as command line flags setting global state, as if the command line tools were the only use of LLVM. Very sad.