LLVM.General.ExecutionEngine

Description

Tools for JIT execution

Documentation

class ExecutionEngine e f | e -> f where

http://llvm.org/doxygen/classllvm_1_1ExecutionEngine.html

Methods

withModuleInEngine :: e -> Module -> (ExecutableModule e -> IO a) -> IO a

getFunction :: ExecutableModule e -> Name -> IO (Maybe f)

data ExecutableModule e

a ExecutableModule e represents a Module which is currently in an ExecutionEngine, and so the functions of which may be executed.

data JIT

http://llvm.org/doxygen/classllvm_1_1JIT.html

withJIT

Arguments

:: Context

-> Word
optimization level

-> (JIT -> IO a)

-> IO a

bracket the creation and destruction of a JIT

data MCJIT

http://llvm.org/doxygen/classllvm_1_1MCJIT.html http://blog.llvm.org/2010/04/intro-to-llvm-mc-project.html N.B. - the LLVM MCJIT does not current support adding multiple modules to any one instance of the MCJIT.

withMCJIT

Arguments

:: Context

-> Maybe Word
optimization level

-> Maybe Model

-> Maybe Bool
True to disable frame pointer elimination

-> Maybe Bool
True to enable fast instruction selection -> Maybe MemoryManager -- llvm-general doesn't support this yet

-> (MCJIT -> IO a)

-> IO a

bracket the creation and destruction of an MCJIT