LLVM.General.Module

Description

A Module holds a C++ LLVM IR module. Modules may be converted to or from strings or Haskell ASTs, or added to an ExecutionEngine and so JIT compiled to get function pointers.

Documentation

data Module

http://llvm.org/doxygen/classllvm_1_1Module.html

withModuleFromAST :: Context -> Module -> (Module -> IO a) -> ErrorT String IO a

Build an LLVM.General.Module from a LLVM.General.AST.Module - i.e. lower an AST from Haskell into C++ objects.

moduleAST :: Module -> IO Module

Get an LLVM.General.AST.Module from a LLVM.General.Module - i.e. raise C++ objects into an Haskell AST.

withModuleFromString :: Context -> String -> (Module -> IO a) -> ErrorT (Either String Diagnostic) IO a

parse Module from LLVM assembly

moduleString :: Module -> IO String

generate LLVM assembly from a Module

writeBitcodeToFile :: FilePath -> Module -> ErrorT String IO ()

generate LLVM bitcode from a Module

writeAssemblyToFile :: TargetMachine -> FilePath -> Module -> ErrorT String IO ()

write target-specific assembly directly into a file

writeObjectToFile :: TargetMachine -> FilePath -> Module -> ErrorT String IO ()

write target-specific object code directly into a file

linkModules

Arguments

:: Bool
True to leave the right module unmodified, False to cannibalize it (for efficiency's sake).

-> Module
The module into which to link

-> Module
The module to link into the other (and cannibalize or not)

-> ErrorT String IO ()

link LLVM modules - move or copy parts of a source module into a destination module. Note that this operation is not commutative - not only concretely (e.g. the destination module is modified, becoming the result) but abstractly (e.g. unused private globals in the source module do not appear in the result, but similar globals in the destination remain).