LLVM.General.Transforms

Description

This module provides an enumeration of the various transformation (e.g. optimization) passes provided by LLVM. They can be used to create a PassManager to, in turn, run the passes on Modules. If you don't know what passes you want, consider instead using CuratedPassSetSpec.

Documentation

data Pass

http://llvm.org/docs/Passes.html#transform-passes A few passes can make use of information in a TargetLowering if one is provided to createPassManager. http://llvm.org/doxygen/classllvm_1_1Pass.html

Constructors

AggressiveDeadCodeElimination

BlockPlacement

BreakCriticalEdges

CodeGenPrepare
can use a TargetLowering

ConstantPropagation

CorrelatedValuePropagation

DeadCodeElimination

DeadInstructionElimination

DeadStoreElimination

DemoteRegisterToMemory

EarlyCommonSubexpressionElimination

GlobalValueNumbering

Fields

noLoads :: Bool

InductionVariableSimplify

InstructionCombining

JumpThreading

LoopClosedSingleStaticAssignment

LoopInvariantCodeMotion

LoopDeletion

LoopIdiom

LoopInstructionSimplify

LoopRotate

LoopStrengthReduce

LoopUnroll

Fields

loopUnrollThreshold :: Maybe Word

count :: Maybe Word

allowPartial :: Maybe Bool

LoopUnswitch

Fields

optimizeForSize :: Bool

LowerAtomic

LowerInvoke
can use a TargetLowering

Fields

useExpensiveExceptionHandlingSupport :: Bool

LowerSwitch

LowerExpectIntrinsic

MemcpyOptimization

PromoteMemoryToRegister

Reassociate

ScalarReplacementOfAggregates

Fields

requiresDominatorTree :: Bool

OldScalarReplacementOfAggregates

Fields

oldScalarReplacementOfAggregatesThreshold :: Maybe Word

useDominatorTree :: Bool

structMemberThreshold :: Maybe Word

arrayElementThreshold :: Maybe Word

scalarLoadThreshold :: Maybe Word

SparseConditionalConstantPropagation

SimplifyLibCalls

SimplifyControlFlowGraph

Sinking

TailCallElimination

AlwaysInline

Fields

insertLifetime :: Bool

ArgumentPromotion

ConstantMerge

FunctionAttributes

FunctionInlining

Fields

functionInliningThreshold :: Word

GlobalDeadCodeElimination

InternalizeFunctions

Fields

exportList :: [String]

InterproceduralConstantPropagation

InterproceduralSparseConditionalConstantPropagation

MergeFunctions

PartialInlining

PruneExceptionHandling

StripDeadDebugInfo

StripDebugDeclare

StripNonDebugSymbols

StripSymbols

Fields

onlyDebugInfo :: Bool

BasicBlockVectorize

Fields

vectorBits :: Word

vectorizeBools :: Bool

vectorizeInts :: Bool

vectorizeFloats :: Bool

vectorizePointers :: Bool

vectorizeCasts :: Bool

vectorizeMath :: Bool

vectorizeFusedMultiplyAdd :: Bool

vectorizeSelect :: Bool

vectorizeCmp :: Bool

vectorizeGetElementPtr :: Bool

vectorizeMemoryOperations :: Bool

alignedOnly :: Bool

requiredChainDepth :: Word

searchLimit :: Word

maxCandidatePairsForCycleCheck :: Word

splatBreaksChain :: Bool

maxInstructions :: Word

maxIterations :: Word

powerOfTwoLengthsOnly :: Bool

noMemoryOperationBoost :: Bool

fastDependencyAnalysis :: Bool

LoopVectorize

EdgeProfiler

OptimalEdgeProfiler

PathProfiler

GCOVProfiler

Fields

emitNotes :: Bool

emitData :: Bool

version :: GCOVVersion

useCfgChecksum :: Bool

noRedZone :: Bool

functionNamesInData :: Bool

AddressSanitizer

Fields

checkInitOrder :: Bool

checkUseAfterReturn :: Bool

checkLifetime :: Bool

blackListFile :: Maybe FilePath

zeroBaseShadow :: Bool

AddressSanitizerModule

Fields

checkInitOrder :: Bool

blackListFile :: Maybe FilePath

zeroBaseShadow :: Bool

MemorySanitizer

Fields

trackOrigins :: Bool

blackListFile :: Maybe FilePath

ThreadSanitizer

Fields

blackListFile :: Maybe FilePath

BoundsChecking

defaultVectorizeBasicBlocks :: Pass

Defaults for the BasicBlockVectorize pass - copied from the C++ code to keep these defaults constant. (The C++ defaults are modifiable through global objects used for command-line processing, in a design apparently oblivious to uses of LLVM besides the standard command-line tools).

newtype GCOVVersion

See http://gcc.gnu.org/viewcvs/gcc/trunk/gcc/gcov-io.h?view=markup.

Constructors

GCOVVersion String

defaultGCOVProfiler :: Pass

Defaults for GCOVProfiler.

defaultAddressSanitizer :: Pass

Defaults for AddressSanitizer.

defaultAddressSanitizerModule :: Pass

Defaults for AddressSanitizerModule.

defaultMemorySanitizer :: Pass

Defaults for MemorySanitizer.

defaultThreadSanitizer :: Pass

Defaults for ThreadSanitizer.